+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - In app browser is now switched over to use custom tabs - contribution from @micahmo
 - Adjusted theming options to show a modal rather than toggles to reduce confusion - contribution from @coslu
 - Added ability to share post, external link, or media from post share button - contribution from @micahmo
+- Added ability to disable post and comment swipe actions separately
 
 ### Changed
 - Adjusted subscription styling to be more consistent - contribution from @micahmo